wip separated BundleList from BundleMap
