+++ packages/marlowe-object/src/reference.ts
@@ -1,7 +1,7 @@
 import * as t from "io-ts/lib/index.js";
 
 /**
- * A label for one of the {@link ObjectType}
+ * A label for one of the {@link @marlowe.io/marlowe-object.bundle-list.ObjectType}
  * @category Object
  */
 export type Label = string;
@@ -9,7 +9,7 @@ export type Label = string;
 export const LabelGuard: t.Type<Label> = t.string;
 
 /**
- * A label for one of the {@link ObjectType}
+ * A label for one of the {@link @marlowe.io/marlowe-object.bundle-list.ObjectType}
  * @category Object
  */
 export type ContractSourceId = string;
@@ -17,7 +17,7 @@ export type ContractSourceId = string;
 export const ContractSourceIdGuard: t.Type<ContractSourceId> = t.string;
 
 /**
- * A reference to an {@link ObjectType}.
+ * A reference to an {@link @marlowe.io/marlowe-object.bundle-list.ObjectType}.
  * @category Object
  */
 export interface Reference {
